+++ android/src/com/frostwire/android/gui/transfers/InvalidBittorrentDownload.java
@@ -95,7 +95,7 @@ final class InvalidBittorrentDownload implements BittorrentDownload, InvalidTran
     }
 
     @Override
-    public long getSize() {
+    public double getSize() {
         return 0;
     }
